Crate narwhal_config

Re-exports

pub use committee::*;

Modules

committee

utils

Structs

AnemoParameters

ChainIdentifier

NetworkAdminServerParameters

Parameters

Holds all the node properties. An example is provided to showcase the usage and deserialization from a json file. To define a Duration on the property file can use either milliseconds or seconds (e.x 5s, 10ms , 2000ms).

PrometheusMetricsParameters

WorkerCache

WorkerIndex

WorkerInfo

Enums

CommitteeUpdateError

ConfigError

Traits

Export

Import

Type Aliases

Epoch

The epoch number.

Stake

WorkerId